+++ src/BusinessLogic/SeQuraAPI/Order/OrderProxy.php
@@ -17,7 +17,6 @@ use SeQura\Core\BusinessLogic\SeQuraAPI\Order\Requests\CreateOrderHttpRequest;
 use SeQura\Core\BusinessLogic\SeQuraAPI\Order\Requests\GetAvailablePaymentMethodsHttpRequest;
 use SeQura\Core\BusinessLogic\SeQuraAPI\Order\Requests\GetFormHttpRequest;
 use SeQura\Core\BusinessLogic\SeQuraAPI\Order\Requests\AcknowledgeOrderHttpRequest;
-use SeQura\Core\BusinessLogic\SeQuraAPI\Order\Requests\UpdateOrderCartsHttpRequest;
 use SeQura\Core\BusinessLogic\SeQuraAPI\Order\Requests\UpdateOrderHttpRequest;
 
 /**
@@ -86,14 +85,6 @@ class OrderProxy extends AuthorizedProxy implements OrderProxyInterface
                 $request->getMerchantReference()->getOrderRef1(),
                 $request)
         )->isSuccessful();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function updateOrderCarts(string $id, UpdateOrderRequest $request): bool
-    {
-        return $this->put(new UpdateOrderCartsHttpRequest($id, $request))->isSuccessful();
     }
 
     /**